address PR comments, remove unused imports, move perf test utility to test
