+++ src/test/java/org/apache/platypus/server/grpc/GrpcServer.java
@@ -1,6 +1,5 @@
 package org.apache.platypus.server.grpc;
 
-import com.amazonaws.services.s3.AmazonS3;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.util.JsonFormat;
 import io.grpc.ManagedChannel;
@@ -14,8 +13,6 @@ import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
 import org.apache.platypus.server.luceneserver.GlobalState;
 import org.apache.platypus.server.utils.Archiver;
-import org.apache.platypus.server.utils.ArchiverImpl;
-import org.apache.platypus.server.utils.TarImpl;
 import org.junit.rules.TemporaryFolder;
 
 import java.io.IOException;
